Inital commit
The init function will run when the page loads up. storing collection of the reviewing stars. defined the event listeners for the page. Looping through the collection of the star images and changed the cursors to pointer. Runs the updateCount function in response to the keyup event. the function is to color the star when user moves the mouse over the stars. storing the alt object in the variable star number. When the mouse moves over the stars the stars will light up. The loop has the first star highlighted but can't go anymore than 5. The purpose of this function is to unlighted the stars when the mouse goes back from the last star decreasing to the first star.  Keeps the total amount of characters the user has typed in the box on the page.